moved PlotSlices
